# Add `frule` and `rrule` for `bsplinebasisall`