--- a/test/test_ChainRules.jl
+++ b/test/test_ChainRules.jl
@@ -1,6 +1,7 @@
 @testset "ChainRules" begin
     k = KnotVector(rand(12))
-    P = BSplineSpace{2}(k)
+    p = 2
+    P = BSplineSpace{p}(k)
     @testset "bsplinebasis" begin
         for _ in 1:10
             t = rand(domain(P))
@@ -11,6 +12,10 @@
                 test_rrule(bsplinebasis₊₀, P, i, t)
                 test_frule(bsplinebasis₋₀, P, i, t)
                 test_rrule(bsplinebasis₋₀, P, i, t)
+            end
+            for i in 1:length(k)-2p-1
+                test_frule(bsplinebasisall, P, i, t)
+                test_rrule(bsplinebasisall, P, i, t)
             end
         end
     end