This PR fixes #265.