@hyrodium 2022-09-08
**Before this PR**
```julia
julia> using Zygote, BasicBSpline, BenchmarkTools, Random
[ Info: Precompiling BasicBSpline [4c5d9882-2acf-4ea4-9e48-968fd4518195]

julia> Random.seed!(42)
TaskLocalRNG()

julia> P = BSplineSpace{3}(KnotVector(rand(30)))
BSplineSpace{3, Float64}(KnotVector([0.02698475249996979, 0.030185550039208087, 0.05529822631508752, 0.07933388944734077, 0.09443144857141617, 0.16643864408566544, 0.1735745757945074, 0.18342915487003242, 0.2112283719951349, 0.25858546995315457, 0.32166161915780656, 0.3475083889757653, 0.3723466293304377, 0.3906633890864917, 0.4339142195043123, 0.48302213696845187, 0.49205817172341115, 0.5270150071089016, 0.5447580773835046, 0.6226472050312687, 0.6349805161807821, 0.6599599239949302, 0.6637579955379869, 0.7219827350406843, 0.802762551279973, 0.8120185311491867, 0.837335454222923, 0.9786385378290976, 0.9807576556964709, 0.9893116874624179]))

julia> n = dim(P)
26

julia> a = rand(n,n);

julia> M = BSplineManifold(a,(P,P));

julia> gradient(M, 0.5, 0.4)
ERROR: Need an adjoint for constructor StaticArraysCore.SVector{4, Float64}. Gradient is of type StaticArraysCore.SizedVector{4, Float64, Vector{Float64}}
Stacktrace:
  [1] error(s::String)
    @ Base ./error.jl:35
  [2] (::Zygote.Jnew{StaticArraysCore.SVector{4, Float64}, Nothing, false})(Δ::StaticArraysCore.SizedVector{4, Float64, Vector{Float64}})
    @ Zygote ~/.julia/packages/Zygote/DRjAT/src/lib/lib.jl:327
  [3] (::Zygote.var"#1948#back#224"{Zygote.Jnew{StaticArraysCore.SVector{4, Float64}, Nothing, false}})(Δ::StaticArraysCore.SizedVector{4, Float64, Vector{Float64}})
    @ Zygote ~/.julia/packages/ZygoteRules/AIbCs/src/adjoint.jl:67
  [4] Pullback
    @ ~/.julia/packages/StaticArraysCore/HoT7O/src/StaticArraysCore.jl:107 [inlined]
  [5] (::typeof(∂(StaticArraysCore.SVector{4, Float64})))(Δ::StaticArraysCore.SizedVector{4, Float64, Vector{Float64}})
    @ Zygote ~/.julia/packages/Zygote/DRjAT/src/compiler/interface2.jl:0
  [6] Pullback
    @ ~/.julia/dev/StaticArrays/src/convert.jl:163 [inlined]
  [7] macro expansion
    @ ~/.julia/dev/BasicBSpline/src/_BSplineBasis.jl:225 [inlined]
  [8] Pullback
    @ ~/.julia/dev/BasicBSpline/src/_BSplineBasis.jl:225 [inlined]
  [9] macro expansion
    @ ~/.julia/dev/BasicBSpline/src/_BSplineManifold.jl:77 [inlined]
 [10] Pullback
    @ ~/.julia/dev/BasicBSpline/src/_BSplineManifold.jl:77 [inlined]
 [11] (::typeof(∂(unsafe_mapping)))(Δ::Float64)
    @ Zygote ~/.julia/packages/Zygote/DRjAT/src/compiler/interface2.jl:0
 [12] macro expansion
    @ ~/.julia/dev/BasicBSpline/src/_BSplineManifold.jl:117 [inlined]
 [13] Pullback
    @ ~/.julia/dev/BasicBSpline/src/_BSplineManifold.jl:117 [inlined]
 [14] (::Zygote.var"#60#61"{typeof(∂(λ))})(Δ::Float64)
    @ Zygote ~/.julia/packages/Zygote/DRjAT/src/compiler/interface.jl:45
 [15] gradient(::BSplineManifold{2, (3, 3), Float64, Tuple{BSplineSpace{3, Float64}, BSplineSpace{3, Float64}}}, ::Float64, ::Vararg{Float64})
    @ Zygote ~/.julia/packages/Zygote/DRjAT/src/compiler/interface.jl:97
 [16] top-level scope
    @ REPL[7]:1
```

**After this PR**
```julia
julia> using Zygote, BasicBSpline, BenchmarkTools, Random

julia> P = BSplineSpace{3}(KnotVector(rand(30)))
BSplineSpace{3, Float64}(KnotVector([0.044385887200565466, 0.053038461426597294, 0.06105412686653566, 0.10989095341196053, 0.11718183382339398, 0.1323012590269842, 0.14020729030461887, 0.15076527899706338, 0.1614427641329753, 0.18445114775379545, 0.21027829387231467, 0.22140887556832733, 0.24462184800247666, 0.32291201376826717, 0.35322759636865053, 0.3638440743450353, 0.5461058547575252, 0.5911473148029991, 0.652027330460927, 0.7895751604040312, 0.7976204560035017, 0.7989935638643737, 0.8094253458660154, 0.8441220355503626, 0.8473439222170284, 0.8604651871674578, 0.8850818739692998, 0.9368904583299719, 0.9667909751237374, 0.9871107683889789]))

julia> using Zygote, BasicBSpline, BenchmarkTools, Random

julia> Random.seed!(42)
TaskLocalRNG()

julia> P = BSplineSpace{3}(KnotVector(rand(30)))
BSplineSpace{3, Float64}(KnotVector([0.02698475249996979, 0.030185550039208087, 0.05529822631508752, 0.07933388944734077, 0.09443144857141617, 0.16643864408566544, 0.1735745757945074, 0.18342915487003242, 0.2112283719951349, 0.25858546995315457, 0.32166161915780656, 0.3475083889757653, 0.3723466293304377, 0.3906633890864917, 0.4339142195043123, 0.48302213696845187, 0.49205817172341115, 0.5270150071089016, 0.5447580773835046, 0.6226472050312687, 0.6349805161807821, 0.6599599239949302, 0.6637579955379869, 0.7219827350406843, 0.802762551279973, 0.8120185311491867, 0.837335454222923, 0.9786385378290976, 0.9807576556964709, 0.9893116874624179]))

julia> n = dim(P)
26

julia> a = rand(n,n);

julia> M = BSplineManifold(a,(P,P));

julia> gradient(M, 0.5, 0.4)
(-3.7331250515839387, 2.813433752317371)

julia> @benchmark gradient(M, 0.5, 0.4)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  129.736 μs …   6.399 ms  ┊ GC (min … max): 0.00% … 92.22%
 Time  (median):     134.560 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   144.093 μs ± 210.939 μs  ┊ GC (mean ± σ):  5.03% ±  3.36%

       ▁▄▆█▅▂                                                    
  ▁▂▂▄▆██████▇▅▄▃▃▃▃▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  130 μs           Histogram: frequency by time          162 μs <

 Memory estimate: 62.45 KiB, allocs estimate: 1073.

julia> @benchmark gradient($M, 0.5, 0.4)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  128.693 μs …   6.351 ms  ┊ GC (min … max): 0.00% … 92.97%
 Time  (median):     134.986 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   144.185 μs ± 223.955 μs  ┊ GC (mean ± σ):  5.57% ±  3.50%

          ▁▄▆█▇▇▆▃                                               
  ▂▂▂▂▃▄▅██████████▇▅▅▅▄▄▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂ ▃
  129 μs           Histogram: frequency by time          158 μs <

 Memory estimate: 62.48 KiB, allocs estimate: 1074.
```
The performance on speed seems able to be improved, but avoiding error is enough in this PR.